feat: pull all react flow stuff out of ViewApp into ViewChart
